DON-1204 Switch campaign detail page map to a realistic position and real data
