+++ src/app/campaign-details/campaign-details.component.html
@@ -34,15 +34,6 @@
     </biggive-formatted-text>
   </div>
 
-  <div
-    tabindex="-1"
-    id="map"
-    style="height: 400px; width: 100%; margin-top: 2rem; margin-bottom: 2rem; z-index: 0"
-  ></div>
-  <div tabindex="0" class="visually-hidden" aria-live="polite">
-    <p>{{ impactRegions }}</p>
-  </div>
-
   <div class="c-details-container">
     <div class="b-primary-column">
       <div class="banner-container">
