+++ src/app/campaign-details/campaign-details.component.ts
@@ -1,16 +1,5 @@
 import { DatePipe, isPlatformBrowser, Location, AsyncPipe, CurrencyPipe } from '@angular/common';
-import { firstValueFrom } from 'rxjs';
-import {
-  Component,
-  OnDestroy,
-  OnInit,
-  AfterViewInit,
-  PLATFORM_ID,
-  ViewEncapsulation,
-  inject,
-  input,
-  effect,
-} from '@angular/core';
+import { Component, OnDestroy, OnInit, PLATFORM_ID, ViewEncapsulation, inject, input, effect } from '@angular/core';
 import { HttpClient } from '@angular/common/http';
 import { DomSanitizer, SafeResourceUrl } from '@angular/platform-browser';
 import { ActivatedRoute, Params, Router, RouterLink } from '@angular/router';
@@ -37,8 +26,6 @@ import { CampaignInfoComponent } from '../campaign-info/campaign-info.component'
 import { MatTabGroup, MatTab } from '@angular/material/tabs';
 import { OptimisedImagePipe } from '../optimised-image.pipe';
 import { MatSnackBar } from '@angular/material/snack-bar';
-import { GeoJSON, Map, Rectangle } from 'leaflet';
-import type { FeatureCollection, Feature, Geometry, GeoJsonProperties } from 'geojson';
 
 @Component({
   // https://stackoverflow.com/questions/45940965/angular-material-customize-tab
@@ -68,7 +55,7 @@ import type { FeatureCollection, Feature, Geometry, GeoJsonProperties } from 'ge
     OptimisedImagePipe,
   ],
 })
-export class CampaignDetailsComponent implements OnInit, OnDestroy, AfterViewInit {
+export class CampaignDetailsComponent implements OnInit, OnDestroy {
   isEarlyPreview = input(false);
   private datePipe = inject(DatePipe);
   private http = inject(HttpClient);
@@ -91,88 +78,10 @@ export class CampaignDetailsComponent implements OnInit, OnDestroy, AfterViewIni
   currencyPipeDigitsInfo = currencyPipeDigitsInfo;
 
   private timer: number | NodeJS.Timeout | undefined; // State update setTimeout reference, for client side when donations open soon
-  /** @var List for accessible read-outs **/
-  protected impactRegions: string = '';
 
   ngOnInit() {
     this.campaign = this.route.snapshot.data.campaign;
     this.setSecondaryProps(this.campaign);
-  }
-
-  async ngAfterViewInit() {
-    if (!isPlatformBrowser(this.platformId)) {
-      return;
-    }
-
-    const map = new Map('map', {
-      minZoom: 4,
-      maxZoom: 12,
-    }).setView([51.505, -0.09], 4);
-
-    const constantLayerStyle = {
-      fillColor: '#c9fdd4', // Light minty green for (not project-relevant) land; we'll use BG green to highlight.
-      fillOpacity: 1,
-      color: '#999999', // Grey outline
-      weight: 1,
-    };
-
-    // Blue background – sea / areas where we don't do geo highlighting.
-    new Rectangle(
-      [
-        [-90, -180],
-        [90, 180],
-      ],
-      {
-        fillColor: '#dddddd',
-        fillOpacity: 1,
-        stroke: false,
-        interactive: false,
-      },
-    ).addTo(map);
-
-    const nationData = await firstValueFrom(this.http.get('../../assets/map/nations.geojson'));
-    new GeoJSON(nationData, {
-      attribution:
-        'Boundaries &copy; <a href="https://www.nationalarchives.gov.uk/doc/open-government-licence/version/3/">Crown copyright</a>',
-      style: () => constantLayerStyle,
-    }).addTo(map);
-
-    const englandRegionsData = await firstValueFrom(this.http.get('../../assets/map/englandRegions.geojson'));
-    new GeoJSON(englandRegionsData, { style: () => constantLayerStyle }).addTo(map);
-
-    const localAuthorityData = await firstValueFrom(
-      this.http.get<FeatureCollection>('../../assets/map/localAuthorities.geojson'),
-    );
-    const demoHighlights = ['E09000012', 'E09000028', 'E12000008']; // Should be Hackney, Bermondsey, South East England
-    const matchedRegions: string[] = [];
-
-    // Build a layer with just project-relevant locations and a list of their names
-    const highlightAreas = await this.getHighlightedFeatures(demoHighlights);
-
-    const projectLayer = new GeoJSON(highlightAreas, {
-      style: () => ({
-        fillColor: '#2c089b',
-        fillOpacity: 0.4,
-        color: '#2c089b',
-        weight: 1.5,
-      }),
-      // No leaflet Layer type I can find, so:
-      // eslint-disable-next-line @typescript-eslint/no-explicit-any
-      onEachFeature: (feature: Feature<Geometry, GeoJsonProperties>, layer: any) => {
-        if (feature.properties && feature.properties.name) {
-          layer.bindPopup(feature.properties.name);
-          // Check for unique to avoid duplicates from e.g. unitary authorities that are in 2 sources.
-          if (!matchedRegions.includes(feature.properties.name)) {
-            matchedRegions.push(feature.properties.name);
-          }
-        }
-      },
-    }).addTo(map);
-
-    // Want to change this in one go to encourage screen readers to politely announce a change just once.
-    this.impactRegions = `UK impact is in ${matchedRegions.join(', ')}`; // No 'and' for now, think it's enough to encourage pauses.
-
-    map.fitBounds(projectLayer.getBounds());
   }
 
   constructor() {
@@ -278,37 +187,5 @@ export class CampaignDetailsComponent implements OnInit, OnDestroy, AfterViewIni
     }
 
     return this.campaign.summary.replace(/\n{2,}/g, '\n').replace(/\n/g, '\n\n');
-  }
-
-  /**
-   * Checks all 4 layers for a region code match, returning every matching feature with the best
-   * readable place name in a property `name`.
-   */
-  private async getHighlightedFeatures(regionCodes: string[]): Promise<Array<Feature<Geometry, GeoJsonProperties>>> {
-    const layers = [
-      { path: '../../assets/map/localAuthorities.geojson', codeField: 'LAD25CD', nameField: 'LAD25NM' },
-      { path: '../../assets/map/counties.geojson', codeField: 'CTYUA25CD', nameField: 'CTYUA25NM' },
-      { path: '../../assets/map/englandRegions.geojson', codeField: 'RGN25CD', nameField: 'RGN25NM' },
-      { path: '../../assets/map/nations.geojson', codeField: 'CTRY25CD', nameField: 'CTRY25NM' },
-    ];
-
-    const fetchPromises = layers.map(async (layer) => {
-      const data = await firstValueFrom(this.http.get<FeatureCollection>(layer.path));
-
-      return data.features
-        .filter(
-          (feature: Feature<Geometry, GeoJsonProperties>) =>
-            feature.properties && regionCodes.includes(feature.properties[layer.codeField]),
-        )
-        .map((feature: Feature<Geometry, GeoJsonProperties>) => {
-          if (feature.properties) {
-            feature.properties['name'] = feature.properties[layer.nameField];
-          }
-          return feature;
-        });
-    });
-
-    const results = await Promise.all(fetchPromises);
-    return results.flat();
   }
 }
