+++ src/app/campaign-info/campaign-info.component.html
@@ -68,6 +68,16 @@
           </li>
         }
       </ul>
+      @if (hasRegions) {
+        <div
+          tabindex="-1"
+          #mapElement
+          style="height: 400px; width: 100%; margin-top: 2rem; margin-bottom: 2rem; z-index: 0"
+        ></div>
+        <div tabindex="0" class="visually-hidden" aria-live="polite">
+          <p>{{ impactRegions }}</p>
+        </div>
+      }
     }
   </div>
 </div>
